+++ src/ODR/AdminBundle/Command/RecacheRecordCommand.php
@@ -63,7 +63,8 @@ $output->writeln($e->getMessage());
                 // Delete the job so the queue won't hang, in theory
                 $pheanstalk->delete($job);
             }
-
+/*
+            // TODO - move this check inside WorkerController?  since commands can't have database connections...
             // Check to see if there's any datafields that need migrating
             while (true) {
                 try {
@@ -80,7 +81,7 @@ $output->writeln($e->getMessage());
                     break;
                 }
             }
-
+*/
 
             try {
                 // Get Job Data
@@ -98,7 +99,12 @@ $output->writeln($e->getMessage());
 $output->writeln($data->url);
 
                 // Create the required parameters to send
-                $parameters = array('datarecord_id' => $data->datarecord_id, 'api_key' => $data->api_key, 'scheduled_at' => $data->scheduled_at);
+                $parameters = array(
+                    'tracked_job_id' => $data->tracked_job_id,
+                    'datarecord_id' => $data->datarecord_id,
+                    'api_key' => $data->api_key,
+                    'scheduled_at' => $data->scheduled_at
+                );
 
                 // Set the options for the POST request
                 curl_setopt_array($ch, array(
@@ -114,6 +120,7 @@ $output->writeln($data->url);
                 );
 
                 // Send the request
+                $delete_job = true;
                 if( ! $ret = curl_exec($ch)) {
                     if ( curl_errno($ch) == 28 ) {
 $output->writeln('timeout detected, deleting job and sleeping for 5 minutes...');
@@ -131,10 +138,16 @@ $logger->err('RecacheRecordCommand.php: timeout detected, deleting job and sleep
                     // Do things with the response returned by the controller?
                     $result = json_decode($ret);
                     if ( isset($result->r) && isset($result->d) ) {
-                        if ( $result->r == 0 )
+                        if ( $result->r == 0 ) {
                             $output->writeln( $result->d );
-                        else
+                        }
+                        else if ( $result->r == 1 ) {
                             throw new \Exception( $result->d );
+                        }
+                        else if ( $result->r == 2 ) {
+                            $output->writeln( $result->d );
+                            $delete_job = false;
+                        }
                     }
                     else {
                         // Should always be a json return...
@@ -146,7 +159,10 @@ $logger->err('RecacheRecordCommand.php: timeout detected, deleting job and sleep
                     curl_close($ch);
 
                     // Dealt with the job
-                    $pheanstalk->delete($job);
+                    if ($delete_job)
+                        $pheanstalk->delete($job);
+                    else
+                        $pheanstalk->release($job, 2048, 10);   // release job back into queue as lower priority, on a 10 second delay
                 }
 
                 // Sleep for a bit
